rag-pipeline/requirements.txt
# beautifulsoup4==4.12.3
# requests==2.31.0
# cohere>=5.13.0,<6.0.0
# qdrant-client==1.12.0
# python-dotenv==1.0.1
# lxml==5.1.0
# html2text==2024.2.26


beautifulsoup4==4.12.3
requests==2.31.0
cohere>=5.13.0,<6.0.0
qdrant-client==1.12.0
python-dotenv==1.0.1
html2text==2024.2.26
